feat: EF move to embeddings sub-package
