+++ README.md
@@ -174,7 +174,7 @@ curl http://localhost:11434/api/embeddings -d '{\n  "model": "llama2",\n  "promp
 
 ### Example Cohere Embedding Function
 
-In this example we rely on `tech.amikos.chromadb.CohereEmbeddingFunction` to generate embeddings for our documents.
+In this example we rely on `tech.amikos.chromadb.embeddings.cohere.CohereEmbeddingFunction` to generate embeddings for our documents.
 
 | **Important**: Ensure you have `COHERE_API_KEY` environment variable set
 
@@ -183,32 +183,33 @@ package tech.amikos;
 
 import tech.amikos.chromadb.*;
 import tech.amikos.chromadb.Collection;
+import tech.amikos.chromadb.embeddings.cohere.CohereEmbeddingFunction;
 
 import java.util.*;
 
 public class Main {
-    public static void main(String[] args) {
-        try {
-            Client client = new Client(System.getenv("CHROMA_URL"));
-            client.reset();
-            String apiKey = System.getenv("COHERE_API_KEY");
-            EmbeddingFunction ef = new CohereEmbeddingFunction(apiKey);
-            Collection collection = client.createCollection("test-collection", null, true, ef);
-            List<Map<String, String>> metadata = new ArrayList<>();
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "scientist");
-            }});
-            metadata.add(new HashMap<String, String>() {{
-                put("type", "spy");
-            }});
-            collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
-            Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
-            System.out.println(qr);
-        } catch (Exception e) {
-            e.printStackTrace();
-            System.out.println(e);
-        }
+  public static void main(String[] args) {
+    try {
+      Client client = new Client(System.getenv("CHROMA_URL"));
+      client.reset();
+      String apiKey = System.getenv("COHERE_API_KEY");
+      EmbeddingFunction ef = new CohereEmbeddingFunction(apiKey);
+      Collection collection = client.createCollection("test-collection", null, true, ef);
+      List<Map<String, String>> metadata = new ArrayList<>();
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "scientist");
+      }});
+      metadata.add(new HashMap<String, String>() {{
+        put("type", "spy");
+      }});
+      collection.add(null, metadata, Arrays.asList("Hello, my name is John. I am a Data Scientist.", "Hello, my name is Bond. I am a Spy."), Arrays.asList("1", "2"));
+      Collection.QueryResponse qr = collection.query(Arrays.asList("Who is the spy"), 10, null, null, null);
+      System.out.println(qr);
+    } catch (Exception e) {
+      e.printStackTrace();
+      System.out.println(e);
     }
+  }
 }
 ```
