+++ src/main/java/tech/amikos/chromadb/embeddings/cohere/CreateEmbeddingResponse.java
@@ -1,4 +1,4 @@
-package tech.amikos.cohere;
+package tech.amikos.chromadb.embeddings.cohere;
 
 import com.google.gson.Gson;
 import com.google.gson.annotations.SerializedName;
